completed activity 2 test CMakeLists.txt works
